fix lib_notary

+++ scripts/test_nn_btc_tx_DB_integrity.py
@@ -38,16 +38,22 @@ def validate_ntx_addr():
         category = row[0]
         notary = row[1]
         txid = row[2]
+
         if notary != "BTC_NTX_ADDR":
             print(f"BTC_NTX_ADDR {txid} IMPROPER NOTARY {notary}")
-            cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
-            conn.commit()
-            print(f"deleted {txid}")
+            delete = input("Delete? (y/n)")
+            if delete in ["Y", "y"]:
+                cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
+                conn.commit()
+                print(f"deleted {txid}")
+
         if category != "NTX":
             print(f"BTC_NTX_ADDR {txid} IMPROPER CATEGORY {category}")
-            cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
-            conn.commit()
-            print(f"deleted {txid}")
+            delete = input("Delete? (y/n)")
+            if delete in ["Y", "y"]:
+                cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
+                conn.commit()
+                print(f"deleted {txid}")
 
 
 def validate_ntx_row_counts():
@@ -56,11 +62,15 @@ def validate_ntx_row_counts():
     for txid in ntx_txids:
         cursor.execute(f"SELECT COUNT(*) FROM nn_btc_tx WHERE txid='{txid}';")
         txid_row_count = cursor.fetchall()[0][0]
+
         if txid_row_count != 14:
             print(f"ntx {txid} row count {txid_row_count}")
-            cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
-            conn.commit()
-            print(f"deleted {txid}")
+            delete = input("Delete? (y/n)")
+
+            if delete in ["Y", "y"]:
+                cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
+                conn.commit()
+                print(f"deleted {txid}")
 
 def validate_split_rows():
     print(f"Validating Split Rows...")
@@ -68,11 +78,15 @@ def validate_split_rows():
     for txid in ntx_txids:
         cursor.execute(f"SELECT COUNT(*) FROM nn_btc_tx WHERE txid='{txid}';")
         txid_row_count = cursor.fetchall()[0][0]
+
         if txid_row_count != 1:
             print(f"split {txid} row count {txid_row_count}")
-            cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
-            conn.commit()
-            print(f"deleted {txid}")
+            delete = input("Delete? (y/n)")
+
+            if delete in ["Y", "y"]:
+                cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
+                conn.commit()
+                print(f"deleted {txid}")
         else:
             cursor.execute(f"SELECT input_index, input_sats, output_index, output_sats, txid FROM nn_btc_tx WHERE txid='{txid}';")
             split_row = cursor.fetchall()[0]
@@ -81,12 +95,16 @@ def validate_split_rows():
             output_index = split_row[2]
             output_sats = split_row[3]
             txid = split_row[4]
+
             for i in [input_index, input_sats, output_index, output_sats]:
                 if i != -99:
                     print(f"Non -99 value for input/output index/sats in txid {txid}")
-                    cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
-                    conn.commit()
-                    print(f"deleted {txid}")
+                    delete = input("Delete? (y/n)")
+
+                    if delete in ["Y", "y"]:
+                        cursor.execute(f"DELETE FROM nn_btc_tx WHERE txid='{txid}';")
+                        conn.commit()
+                        print(f"deleted {txid}")
 
 def validate_notaries_by_address(address, season):
     print(f"Validating Notary Address Rows for {address} {season} {NN_BTC_ADDRESSES_DICT[season][address]}...")
@@ -95,25 +113,27 @@ def validate_notaries_by_address(address, season):
     for row in addr_rows:
         notary = row[0]
         txid = row[1]
+
         if NN_BTC_ADDRESSES_DICT[season][address] != notary:
             print(f"{address} is {NN_BTC_ADDRESSES_DICT[season][address]}, not {notary}")
-            sql = f"UPDATE nn_btc_tx SET notary = '{NN_BTC_ADDRESSES_DICT[season][address]}' WHERE address = '{address}' and txid = '{txid}';"
-            cursor.execute(sql)
-            conn.commit()
-            print(f"Updated {txid} {notary} {address} {season}")
+            update_it = input("Update? (y/n)")
 
+            if update_it in ["Y", "y"]:
+                sql = f"UPDATE nn_btc_tx SET notary = '{NN_BTC_ADDRESSES_DICT[season][address]}' WHERE address = '{address}' and txid = '{txid}';"
+                cursor.execute(sql)
+                conn.commit()
+                print(f"Updated {txid} {notary} {address} {season}")
 
-
-
-    # NTX has 13 vins, 1 vout
-    # no tx has rows with same vin/vout combo
-    # no tx has different categories
-
-    pass
 
 validate_ntx_row_counts()
 validate_split_rows()
 validate_ntx_addr()
 for season in NN_BTC_ADDRESSES_DICT:
-    for address in NN_BTC_ADDRESSES_DICT[season]:
-        validate_notaries_by_address(address, season)
+    if season in ["Season_3", "Season_4"]:
+        for address in NN_BTC_ADDRESSES_DICT[season]:
+            validate_notaries_by_address(address, season)
+
+print(f"Categories in DB:")
+cursor.execute(f"SELECT DISTINCT category FROM nn_btc_tx;")
+cat_rows = cursor.fetchall()
+print(cat_rows)
